[fix]募金登録時にユーザが正常に登録されない問題を修正
